backend/requirements.txt
fastapi
uvicorn
python-multipart
numpy
torch --index-url https://download.pytorch.org/whl/cpu
openai-whisper
